Show assessor details on assessor overview page
We need to include identifying information about each submitted
application so that it can be included in the email notifications that
are sent to the assessor inbox. Therefore, we are adding a new page for
assessors to enter this information.

We display the assessor name and referral ID if they are present. If
not, we show a placeholder.