Format (and sanitise) ACCT details with newlines
The main motivation of this came about to prevent the possibility of
cross-site-scripting (XSS) attacks, as when playing this data back on
the ACCTs page, we weren't sanitising the output, meaning that scripts
could, if not for our Content-Security-Policy, be executed. Our recent
pen-test highlighted this as an area that could be reinforced in case of
the CSP failing for any reason.

We actually get sanitisation for free with the `formatLines` function,
which is useful here as we aren't formatting any paragraphs the user
enters in the input field for ACCTs, so we solve two problems here!

TODO: I'm not sure if we need to add this to the `response` function,
but I'm new to the formbuilder pattern so I'm not totally sure what it's
used for yet...