Format and santise ACCT and behaviour note details
The main motivation of this came about to prevent the possibility of
cross-site-scripting (XSS) attacks, as when playing this data back on
the ACCTs page, we weren't sanitising the output, meaning that scripts
could, if not for our Content-Security-Policy, be executed. Our recent
pen-test highlighted this as an area that could be reinforced in case of
the CSP failing for any reason.

We actually get sanitisation for free with the `formatLines` function,
which is useful here as we aren't formatting any paragraphs the user
enters in the input field for ACCT or behaviour notes, so we solve two
problems here.

My first approach for this was to call `formatLines` in the formbuilder
task list functions, but this meant we were doubly `formatLines`-ing on
the Check your answers page due to the way we loop over all task list
classes and format them, so it was printing out e.g. `<p>` tags around
newlines.

Making `formatLines` a nunjucks environment variable that can be called
in templates felt like the easier option than refactoring a lot of that
behaviour.

Integration tests are a bit heavy-handed here, but I felt it was
important to test this behaviour somewhere as we can't unit test it.