Add `validateReferer` function
This can be used to guard against reflective XSS attacks where the
request header's `referer` field is changed. We use the `referer` value
in some back links to place the user on the page they've come from, but
this can't be relied on.

This checks that the user's referer starts with the URL of the site,
e.g. localhost in the local environment, or the dev, preprod or prod URL
if they're on that environment.

We've opted to fall back on the root path `/` in cases where the referer
isn't from our site, for any reason, so we don't break the user journey.
We can't see this being a huge issue as it's incredibly unlikely this
situation will occur.

I've discovered in integration tests that the `referer` may be
undefined, so we're also defaulting to `/` as a back link in that
scenario.