add new `assessmentsTableRows`
this allows us to use the previously-introduced top level fields for
the person's name, crn and nomsNumber on a submitted application summary.

In the future if we do a similar thing for the referrer's dashboard (which uses the`applicationSummary` type) this can be
re-used and the function `submittedApplicationTableRows` be removed.