Validate referer in statusUpdateController
I've had to add a `resetAllMocks` in a `beforeEach` here as otherwise
the expectation for `validateReferer` not being called fails due to it
being called in the `it` block above. Putting it at top level (outside
the `new` block) causes the API expectation below to fail, so I've opted
to scope it to the `new` test here to keep the number of changes to a
minimum.

We also need to add the `INGRESS_URL` env var to `feature.env` as part
of this, otherwise the test environment can't verify the `referer`
matches the `config.domain` value. Before this change, it's expecting it
to be on port `3000`, though the integration tests are running on port
`3007`.