Introduce date util for generating todays date
As we now include a validation on the HDC license page to check if the
CRD is today or in the future, we can no longer rely on fixed dates in
our integration and e2e tests and need to generate these dates in real
time.

In this commit we introduce a util that will generated todays date and
optionally add a number of months to generate a date in the future.

In regards to the test, it was difficult to test without rewriting much
of the logic contained in the function itself, so testing for the
existence of the keys we expect seemed like a good alternative.