add homepage tile for prison applications for Licence CA's
CAS2 currently is setup for users with the POM
role to be able to create/read applications. We
have since found that the ROLE_POM is given out to
a broad range of users who may not be POMs - The
LICENCE_CA role [1] is more in line with who our
users are, and is also used by HDC services to
perform similar actions that we require. We are
therefore going to add in the LICENCE_CA role and
remove the POM role permissions in stages.

Having done the work on the API to allow read
access to LICENCE_CA[2], this allows the users to
see the prison dashboard tile.

[1]https://dsdmoj.atlassian.net/wiki/spaces/CVL/pages/3540320445/Users+and+Roles
[2] https://github.com/ministryofjustice/hmpps-approved-premises-api/pull/1786