refactor e2e setup step
previously this was signing the POM user up before every test -
this is no longer useful as we have several different users.
I've moved the pom sign in to its own function, and replaced the setup to
sign out a user, as we need to do this before every test to ensure they can
run independently or in any order.

As all the tests have a sign in function now, in the next commit
I will create a shared sign in function for them to call.