use Assessment ID to add Status Updates
Passing the Assessment ID down to the service/client layer
to add new Status Updates and Status Update Details.
We also now get the Status Updates from an Assessment rather than directly on an Application.
We still need the Application ID as we have not yet completely
re-pointed the views/model to get the Application overview view, so
I am passing that ID as a query param.