Fix back buttons from pre-task list tasks
This change ensures that for the confirm eligibility page, no back button will render when it is accessed prior to reaching the task list. This makes sense as continuing from the previous page has completed the irreversible step of creating a new application.

For the confirm consent and HDC pages, the back buttons will point to confirm eligibility and confirm consent, respectively.