e2e test for clicking on an application by a different user
We need to test that our POM and LCA users can view and
add notes to applications on their prison dashboard.

Previously we hardcoded the application ID to be used locally
and in dev environment, because we had a standard
prison for our users across both environments.

Since we now are restricting users to only create applications
for offenders within their own prison, we have discovered that
the example offenders in dev and local have different prisons (LEI for local,
MDI for dev). [1] So our local and dev test users do not have the same prison
any more, and that hardcoded UUID no longer works.

In this commit we use the user's name to find an application by any user
that is not them, thus removing the need for a hardcoded id.

This has the added benefit that it's agnostic to what the seeded data is,
and when we introduce pagination we don't care what page we are on.

[1] https://github.com/ministryofjustice/hmpps-approved-premises-api/pull/1879