Application client function to cancel application
We want to allow the user to cancel in progress
applications.

This commit adds the function for the UI client to
make the PUT request to the API to abandon the
application - while we are using the term 'cancel'
on the UI, the API is using the term 'abandon' to
make it clear that this action is taken on
applications that have not been submitted, and
distinguish it from the assessor status of
'referral cancelled'.