API path to make abandon request to
We want to allow the user to cancel in progress
applications.

This adds the API path the client will shortly be
setup to make the request to cancel - while we are
using the term 'cancel' on the UI, the API is
using the term 'abandon' to make it clear that
this action is taken on applications that have not
been submitted, and distinguish it from the
assessor status of 'referral cancelled'.