Add Cancel link to referrer's dashboard
We want to allow users to cancel in progress applications.

This adds a link to the Cancel form on each row of the dashboard,
next to every in progress application, via the utility method which builds the
table data.

Including the integration and e2e tests for the user journey.