Remove integration test for CYA with previous question schema
This test is not a viable long term solution, because any time we add a new question to the form (as in this PR), we can no longer reach the check your answers page, because the application data (according to the previous schema) is incomplete. This is the desired behaviour and to retain the test as intended we would have to update the “previous schema” application data to include the new question, which is obviously wrong.

Really this needs a different solution to dynamically check which pages and questions we are expecting data for in CYA, but in the absence of that, this test in its current form has served its purpose.