fix cancel application e2e's
Previously this test just chose the last in
progress application to cancel, by looking for the
last `Cancel` anchor tag.

This meant that depending on the state of
applications in the database the name of the
person on the application could have been
anything, and we need to assert that the page is
showing the applicant's name in the header.

This changes the test to create an in progress
application for our standard person, to ensure
there will always be a most recent application for
that person first in the list, and then selects
that application row to cancel it.