Update funding source error condition
We want to capture the possibility that a user will have previously entered ‘both’ on this page, and since tried to access the form with the updated question schema. In this case that this.body.fundingSource is ‘both’, we want the page to error so show the task as still in progress, which guards against the application being submitted with incorrect data. So we update the condition to explicitly check that the answer is either ‘personalSavings’ or ‘benefits’.