Application page keys are filtered out from CYA automatically
When removing a whole page we broke access to the CYA page for applications that were already in progress.

For new applications the latest questions are copied from questions.ts into application.data on creation. For new applications access to the CYA page should always be compatible.

For applications started against an older version of the questions schema we have to tell the CYA page not to try and render those pages and/or their questions.

NB. The changes to questions (rather than tasks) is already handled by this logic[1].

We did have a manual guard against this[2] in the form of an if statement that we needed to remember to manually update. This is obviously fragile and we forgot to do it so we try and make that better for next time.

NB. Tasks can have many Pages but Pages don’t have sub pages. So we can look a the top level task keys to find the page keys we need to compare.

[1] https://github.com/ministryofjustice/hmpps-community-accommodation-tier-2-ui/blob/60554b22eec1f11d2735d551d95356156bf11415/server/utils/checkYourAnswersUtils.ts#L93
[2] https://github.com/ministryofjustice/hmpps-community-accommodation-tier-2-ui/blob/60554b22eec1f11d2735d551d95356156bf11415/server/utils/checkYourAnswersUtils.ts#L44