Add exceptions for data pages
The one exception to our dynamic task/page/question checking system is our ‘data’ pages (ACCT, current offences and historic offences), which we use to create and add to lists for those data types.

These lists are collated and rendered via response methods on their index pages. But because these pages do not have any questions, they would be excluded by removeAnyOldPageKeys. So, we add this explicit exception to prevent that.