update "confirm consent" and "confirm eligibility" legend to be an h2
Visual heading on page is not semantically coded as such

On the Confirm Consent page, there is a prominent text label that reads, "Has Catherine Hettinger given their consent to apply for CAS-2?" While the label has been visually styled to resemble a heading and stand out for sighted users, it has not been semantically coded as a heading level 2 `<h2>` element. As a result, although sighted users benefit from the visual styling to quickly see and understand the significant topics of the page, screen reader users do not receive the same programmatic cues.

This leads to a mismatch between the visual and programmatic experience of the page for users who rely on screen readers to navigate through the page. Moreover, screen reader users often use the headings on a page to quickly identify and jump to the main contents of the page.