Remove Risk factors page from application process
Removed two fields because we are caturing this information already from users.

- "What circumstances are likely to increase risk?"
- "When is risk likely to be greatest?"

POMs will not longer be required to fill in this information for new applications.
Historical data will exist but will eventually be phased out. The reason for not removing/hiding
existing data is because we know that this information will eventually be phased out and do
not want to add conditional statements in the code to hide these two questions from users.