Fix document hierarchy
"Contents" should really be heading level 2 and not level 3 as its currently being used between and `h1` and `h2`. Although its less prominent on the page it is the heading for the sidebar navigation. This fix was also recommended as part of an external accessibility audit which raised this issue.