Fix styling of printed/downloaded applications
The text in the blue banner does not meet contrast ratio requirements when a user prints/downloads.
Instead for print/downloaded media, we need to change the font colour back to black.

I also refactored the existing styling which used a lot of `!important` which not something we should encourage. By being a little more specific there is not real need for forcing those styles.