Fix usage of grid layout and improve printed version
Firstly, the page was not using the grid system correctly, i.e defining the column layout directly after a grid row.

Secondly, there was an accessbility issues of the sidebar nav items wrapping on to new lines and how the browser interrupts that. Instead when the page is printed the sidebar should be full width and positioned at the top of the page as it would if the user was on a mobile.