Make the first nav link focusable for "back to top"
In the sidebar, we have a long list of navigation items and the last one says "Back to top". Currently it sets the viewport to the top of page but does not focus any element. There is no anchor point with an id of `top` which is why the viewport takes the user to the top instead of top of the navigation list.

This commit ensures that the first link in the sidebar will now have an id of `top` which the `Back to top` can go to and focus the user on.