Refactor copy to clipboard js
We shouldn't be hardcoding element id just to bind click events. The refactor uses css classes to find the same buttons and attach click events.

We also do no need tto wait until the DOM content has loaded to run this code as its positioned at the bottom of the DOM so browsers would have parsed it by the time the js is run.