Add feedback when Copy Text button is interacted with
An external accessibility audit, raised an issue with the "Copy text" functionality that we have on the page. It reported that there was no feedback that the action had taken place and lead some to be uncertain that the button has functioned correctly.

What I implemented was to change the Copy text button content from `Copy text` to `Text is copied`. This is change is acknowledged by screenreaders as the button now setup to use `aria-live: assertive`. After 5 seconds the content is reverted back to `Copy text`