Improve e2e tests
There was a flickering spec where the tests expect there to be no other data.

What seems to be happening is that the selectors playwright was using was including results from "Submitted" as they are on the page but are visually hidden. The selector was also including the table headings.