Fix styling of the Rosh summary component
This is an existing bug where the css was relying on markup and the markup did not match, ie using h2 instead of h3. Instead we should be using a class and putting the class on the element that we want to style rather than relying on the nested styling.