Move manual rosh info response to summary response
Now that summary is support two different inputs, OASys and Manual RoSH, it felt like a good case to also keep the response for manual rosh information together instead of in the task itself. To overcome some issues with manual rosh not having a responses, it was decided to return an empty object.